[IMP] documents/saas~16.3: add file_extension field in document
To ease the search of documents by format, we have added the file extension in
the document model allowing search group by and sort on extension.
After a first implementation as a computed field based on mimetype (not
merged), we have opted for a stored field to have a more robust implementation.

The upgrade initializes the file extension field from the related attachment
name if possible otherwise from the mimetype.

See odoo/enterprise#39131

Task-3061869

closes odoo/upgrade#4517

Signed-off-by: Christophe Simonis <chs@odoo.com>