[IMP] point_of_sale/saas~16.3.1.0.1: remove background loading fields
In the related community PR, we get rid of the background loading
for products and partners as it was causing problems to companies
with lots of partners or products. We now load a fixed amount
of products (20000) and partners (100) at the launch of the PoS.
To load more products (or partner), the user will have to scan
it (barcode) or simply search it in the database through
the searchbar and especially the search more button.

See odoo/odoo#118006

closes odoo/upgrade#4531

Signed-off-by: Christophe Simonis <chs@odoo.com>