[IMP] l10n_ke_edi_tremol/saas~16.4: remove hsn
fields, create appropriate replacement taxes

See: https://github.com/odoo/odoo/pull/112013

The HSN fields on the product product and product template are removed
in the above PR. They are replaced with a many2one field on the tax,
which can be set as the default tax on the product, but is independent
of the product. This many2one field references l10n_ke.item.code, a new
model introduced in the same commit, which enumerates the complete set
of KRA item codes. The representation of these codes directly on the
product (in the char field) has been replaced is replaced with a
relationship between the tax and these codes. This helps to address
structural issues surrounding the use of these codes (which are more-so
'KRA Item Codes', as opposed to HS codes).

However, existing users might have already configured appropriate codes
on their products. So this commit adds a post-migrate script that
creates appropriate taxes (with the correct vat amount, item code,
repartition lines, description, etc.) for each different code used by a
product. It then assigns the tax to this product, such that it is a
default tax on the product when the code in the 'l10n_ke_hsn_code' field
matches the code in the 'l10n_ke_item_code' field on the new tax.

Note: A new tax is only generated for the product if the product has
been used in at least one account move. In this case, the tax is
generated for the product on every company that uses the 'ke' chart
template.

The l10n_ke_hsn_code/name fields are only removed from the
product_product/product_template after this.

The respective views for the l10n_ke_hsn_code/name on the
product_product and product_template forms are removed in the
pre-migrate.

Considering the nuances of this behaviour, a test has been included to
ensure that the correct taxes are generated (or alternatively that the
taxes are not generated when they wouldn't be valid).

closes odoo/upgrade#4528

Task-id: 3171523
Related: odoo/odoo#112013
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>