[IMP] helpdesk/saas~16.3: implies new group to helpdesk user group
This commit adds the new group in the one for Helpdesk User if the
auto-assignment feature is enabled to at least one helpdesk team.
The goal of that new group is to be able to display filter
when the feature is enabled in at least one helpdesk team, if no
helpdesk team has the feature enabled then the filter is useless and
so it will be hidden thanks to that group.

task-3052920

Enterprise PR: odoo/enterprise#33835

closes odoo/upgrade#4527

Signed-off-by: Christophe Simonis <chs@odoo.com>