[FIX] util/orm,/mail/saas~11.5: do not remove `is_mail_thread` flag
The flag `is_mail_thread` is used by `mail` module to properly setup
custom manual models. If we remove the flag the manual model won't get
the inherit from `mail.thread`. This causes the field
`message_is_follower_id` to loose it's search method that comes from
`mail` module. Since the field is non-stored, it's then marked as
non-searcheable and this makes any custom view for the custom manual
model using that field in a domain to fail.

Similarly the flags `is_mail_activity` and `is_mail_blacklist` would
also add the inherit.
https://github.com/odoo/odoo/blob/904897d3d6c1d7d2a363c5d822b7ac7da7008d6a/addons/mail/models/ir_model.py#L96-L110

It seems the intention when removing the flag was to avoid issues with
`message_main_attachment_id` column that may be missing. To avoid this
we add the column for all models with the is_mail_thread flag set. Note
that between the three flags this is the only one that exists in v11.0

closes odoo/upgrade#4530

See: odoo/odoo@3c2a92e odoo/upgrade@6d10938
Signed-off-by: Christophe Simonis <chs@odoo.com>