[FIX] project/saas~16.2: set waiting state when task is blocked by another
Before this commit, when a task A has another task B in its dependencies and
task B is not yet done then the state of the task A should be set to `Waiting`
instead of following the value of `kanban_state` field.

This commit fixes the issue by setting `Waiting` state when `is_blocked`
field is true otherwise the state will be set according to the
`kanban_state` field.

Issue found during the upgrade of our prod.

Related PR: odoo/upgrade#4367
Community PR: odoo/odoo#107593
Enterprise PR: odoo/enterprise#35359

closes odoo/upgrade#4536

Signed-off-by: Christophe Simonis <chs@odoo.com>