[FIX] base/0.0.0: handle conflicting groups for fields
Field usage in views is now checked against a combination of groups.
This means that the view must be already present in the view without
groups restriction or with a valid combination of groups depending on
the way it is restricted.

The simple solution is to add the restricted fields without any group to
the view. Adding the field with `invisible="1"` doesn't affect the
rendered output while still allows the same functionality.

There are multiple cases where the group restriction kicks in. We handle
here the three encountered so far but we may need to add more soon.

Part-of: odoo/upgrade#4559