[FIX] util/orm: avoid inconsistencies on MRO
If we mark an unloaded standard model that inherits from mail_thread and
mail_activity as manual we may get the following error:
```
  File "/tmp/tmp7q5cs6ro/migrations/util/orm.py", line 478, in custom_module_field_as_manual
    env.registry.setup_models(env.cr)
  File "/home/odoo/src/odoo/15.0/odoo/modules/registry.py", line 279, in setup_models
    model._prepare_setup()
  File "/home/odoo/src/odoo/15.0/odoo/models.py", line 2969, in _prepare_setup
    cls.__bases__ = cls.__base_classes
TypeError: Cannot create a consistent method resolution
order (MRO) for bases BaseModel, mail.thread, mail.activity.mixin, mail.thread.blacklist, base
```

The issue is that sometimes a standard model is left over by the
uninstall process. In such situation the model is not in the loaded pool
(rightfully) and, before this patch, it was thus assumes as a custom
model. Later when the ORM try to setup models, assuming the leftover
model as manual, it fails with errors on the inheritance spec.

In this patch we avoid the issue by leaving out of the conversion into
manual the models that seem to come from standard modules.

tbg-631

Tested on upg-748426 and upg-750621, in both cases conflicting model was
`crm.lead`:
```
2023-04-18 10:24:40,978 18 WARNING 748426_afu_16.0 odoo.upgrade.util.orm: Leftover models from uninstalled standard modules: ['crm.lead']
```

closes odoo/upgrade#4562

Signed-off-by: Christophe Simonis <chs@odoo.com>