[FIX] mrp/saas~16.3: override workcenter times when capacity is defined
In the current form, when a specific capacity is defined for a
workcenter, the setup/cleanup times defined there are added to the
workcenter's setup/cleanup times.

We instead want for the times defined in the capacity to fully override
the times defined in the workcenter.
Changing this breaks the current behavious, so to keep the data
consistent after upgrade, we add all existing workcenter times in their
related capacities, so that the current configuration stays the same
after upgrade, while being now able to fine tune the specific times.

See odoo/odoo#119212

closes odoo/upgrade#4576

Signed-off-by: Christophe Simonis <chs@odoo.com>