[FIX] util.move_model: move mail.template
In `saas~13.4`, some data and views were renamed and/or moved from
`sale_coupon` to `coupon`.

A commit from the `util.move_model` refactor broke the upgrade because
util.move_model doesn't handle `mail.template`

This resulted in a duplicate in `14.0`, and when upgrading to
`15.0`, where some of the data has again been moved from `coupon` to
`sale_coupon`, the duplicate clashed.

upg-753719

closes odoo/upgrade#4570

Signed-off-by: Christophe Simonis <chs@odoo.com>