[IMP] l10n_nl/saas~16.2: use vat to set the peppol fields
The VAT can be used as the Endpoint with EAS code '9944' for Dutch
partner (see https://docs.peppol.eu/poacc/billing/3.0/codelist/eas/).
This way, the peppol fields will be filled and we will not block the
user when he sends and generates a bis 3 attachment (even if the OIN/KVK
were not filled).

On the other hand, the heuristic used to set the peppol fields
(`EAS_MAPPING`) is not updated because we don't want to set this code
automatically by default since it's less reliable then the OIN/KVK
fields.

see https://github.com/odoo/odoo/pull/120621

closes odoo/upgrade#4591

Signed-off-by: Christophe Simonis <chs@odoo.com>