[FIX] account/16.0: generated missing balance payment term line
payment term used in SO but for which payment term line
is missing and this will raise error when we are going
to create invoice and select those payment term.

issue raised when customer does not have payment term line in
originl database, so after upgrade we set default lines
in standard code but it was not set for existing records during
upgrade.
here is default function to set value
https://github.com/odoo/odoo/blob/04d11bf763068dabedc667266e1a6adf0b5db8c0/addons/account/models/account_payment_term.py#L15

Default lines value exist before 16 version but
this error raised only in Odoo 16 when there was not default line.
because of this change :
https://github.com/odoo/odoo/pull/91490/files#diff-1082eea1e176daccb75f95f20dbfed8f869328635fd041c2f6d6d082f5f05563R16
so that's why we need to fix in Odoo 16.

so to avoid this error we need to insert the value in payment term line
for those payment term does not have any balance lines.

upg-575952
upg-603172
upg-596264
upg-561149

unicorn task id : 3266789

An error message is popping up when trying to create an invoice from a SO:
```
An error has occurred. The move (Draft Invoice (* NUMBER)) is not balanced.
The total of debits equals 0.00 € and the total of credits equals 70.00 €.
You might want to specify a default account on journal "Customer Invoices"
to automatically balance each move.
```

closes odoo/upgrade#4602

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>