[FIX] base/0.0.0: check more children
A primary view may be affected by an extension view of one of its
parents up to its root view. Thus we should try to disable them also when
trying to fix a view. Since any extension view must have some primary
view in it's parent path to root, we could limit this heavier check to
only primary views.

We can also limit the set of children that we disable to extension
children. Any direct primary children of a view won't be involved in
it's final arch. We do not need to extend the children further down
because just disabling the direct children breaks the link.

In a visual way:
```
       P0 -> Root
      / \
    P1   E3
   / \   |  \
  E1 E2  P2 E7
  |
  V -> current view
 / \
E4 P3
|  |
E5 E6
```

The general principle is that when rendering a primary view we apply all
its extension descendants, but not primary ones, and then we apply this
to the rendering of its parent. For example to render `P0` we need to
apply `[E3, E7]` but ignore the subtrees of `P1` and `P2`. Similarly to
render `P1` we need to apply all its extensions `[E1, E2, V, E4, E5]` if
`V` is an extension, otherwise `[E1, E2]`, and then we need to apply the
result to the rendering of `P0` (as before).

The path to root (view with no parent) from `V` is `[E1, P1, P0]`. With
those rules in mind we can see that what actually affects the rendering
of `V` is `[E4, E5, P1, E1, E2, P0, E3, E7]`. Now while trying to fix
`V` it doesn't makes sense to disable an extension view on its path to
the root view. We should instead disable views that are either in the
extension subtree of `V` that is `[E4, E5]`, or in the extension
subtrees of any of its primary ancestors but still not not in the path
to root, i.e. `[E2, E3, E7]`. Now since disabling a view will remove
its subtree we can limit the disabling to `[E4, E2, E7]`. For example
if what is causing the failure of `V` is `E5` we will detect it when
re-enabling `E4` and jumping into fixing it, `E5` is a direct child of
`E4` thus it will be tested there and we jump into fixing it.

closes odoo/upgrade#4507

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>