[FIX] sale_subscription/saas~16.1.1.1: Set sale.order.alert.company_id
The good/bad health logic of sale.order.templates was transferred to
sale.order.alerts. For this, sale.order.alerts are being created for all
sale.order.templates that have a (non-empty) `good_health_domain` or
`bad_health_domain`.

If a sale.order.template has both a good/bad health domain and a non-empty
company_id, then a sale.order.alert will be created with an empty company_id.
This, however, will mark the sale.order.alert as inconsistent since its
(null) company is not a superset of its linked template's (non-null) companies.

We can explicitly set the alert's company_id to be the same as its template's,
like we do with the other fields of the sale.order.alert record to be created;
this way, none of the records will be marked as inconsistent.

closes odoo/upgrade#4623

Signed-off-by: Christophe Simonis <chs@odoo.com>