[IMP] onboarding,appointment/saas~16.3: update panel layout fields
Part of the webclient's redesign is an update of the onboarding panels.
The appointment one uses the onboarding module to store layout
information.
There is no longer a background image nor color defined on the panel
but images and alt text for each step.

See odoo/odoo#116641
See odoo/enterprise#38775

Task-2818586

closes odoo/upgrade#4652

Signed-off-by: Christophe Simonis <chs@odoo.com>