[FIX] account/saas~15.4.1.0: Recompute fields
Summary of Changes

This issue is caused due to recent change in the code to get the dashboard
data from 15.0 to 16.0.
15.0: https://github.com/odoo/odoo/blob/d0aad0bfac8eafdda6e3d748bd164e078f77c99f/addons/account/models/account_journal_dashboard.py#L347
16.0: https://github.com/odoo/odoo/blob/4dee34e56f3e3f90724ead7f80bf0cf959581843/addons/account/models/account_journal_dashboard.py#L398

In the 16.0 instead of using three diffrent query, the code is optimised
to get results from single query.

Issue: The amount_residual field is an compute field which is set 0
in this upgrade script of saas~15.4
Link : https://github.com/odoo/upgrade/blob/be3f442b40f5c8055c04d132bbc5ebd6da50d652/migrations/account/saas~15.4.1.0/pre-migrate.py#L5

And as the field is store true the field is not recomputed untill
some changes is done which is why the dashboard data is computed as 0.
FIX: Recompted amount_residual and amount_residual_signed in the
post migrate of saas~15.4

Reference:
Customer Issue ID: 3280435
Request ID : upg-761709

closes odoo/upgrade#4616

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>