[FIX] crm/saas~16.3: undo mark_as_lost renaming
In odoo/odoo#118494
We replace the server action binding
with a binding of the action window it called.
This renders the action useless.

We remove the renaming of the action added for odoo/odoo#116287 as
it is pointless to rename it when it's not used.

As the action remains functional, we don't remove it to allow users
who use it in their customized flows can keep using it for now.
Although it should be considered deprecated.

task-3272955

closes odoo/upgrade#4640

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>