[FIX] account_asset/16.0: inversion account deferred revenue
With the migration to 16.0, the accounts (account_depreciation_id
and account_depreciation_expense_id) have been inverted for
deferred revenues.
The computation has been fixed to adapt to the fact that the
accounts have been inverted.
But the old deferred revenues (before 16.0) won't work anymore.
This fix will inverse the accounts, so models/revenue created before
16.0 continue to work. We can't change current code, as revenues
created in the last months would not work correctly.

We also need to remove the adaption for deferred revenue in
the computation of depreciation_value as it gives completely
wrong values if people reevaluate existing deferred revenues.
With this, entries before and after upgrade are consistent.

Fixes enterprise https://github.com/odoo/enterprise/pull/28307

opw-3301178

closes odoo/upgrade#4666

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>