[FIX] account/saas~14.4: remove test lock date
The aim of this commit is to prevent an error happening because the lock
date are set by another TestCase.

Context:
Another TestCase is setting all the lock date for today + 15 days.
This wasn't an issue until this commit odoo community commit
e4d1a9e0c770d0c6de3119cd24bb890363e8c459 which doesn't garanty the order
in which test are run.

Before the commit:
The lock date might be set when creating the payment

After the commit:
The lock date is garanteed to not be set

Task-id: None
Part-of: odoo/upgrade#4675