[FIX] account/saas~14.4: account_payment test
The aim of this commit is to fix the test of the account_payment
migation by adding a receivable account to the test data.

Before the commit:
There wasn't any receivable account available to be picked which
resulted in the payment not using any account which goes against an sql
constrains.

After the commit:
The payment gets the previously created receivable account
automatically.

closes odoo/upgrade#4675

Task-id: None
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>