[ADD] util.records: reload translations on record reset
A new section is added to the `util` function `update_record_from_xml`:
if fields are passed via the existing kwarg `reset_translations`
and the module to upgrade from is not in the state `to install` or `to
upgrade`
(e.g. when the function is invoked after the standard upgrade during a
custom upgrade),
the translations are reloaded from their `.po` files.

This is needed in the context of custom modules,
where the standard upgrade has already been performed but
records need to be (fully) reset.

closes odoo/upgrade#4642

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>
Co-authored-by: Christophe Simonis <chs@odoo.com>