[IMP] util.ensure_xmlid_match_record: search harder
Previously, if a matching entry was found in ir_model_data and the record
pointed at by res_id existed, res_id was immediately returned. But, sometimes,
the existing record may not be the right one, i.e. it doesn't match the
"values".
This patch changes the behavior such that in this case, if we can find another
record that actually does match the values, we will update the ir_model_data
entry to reference the id of *that* record instead.

Part-of: odoo/upgrade#4693
Co-authored-by: Christophe Simonis <chs@odoo.com>