[FIX] base/saas~16.2: install new module only when needed
The module 'website_sale' depends on delivery but not on 'stock_delivery'
so there is no need to install the latter, which will force install 'stock'
as well, for those who do not need it.
The condition on force_upgrading the 'stock_delivery' module would still be
needed if it is autoinstalled due to 'stock' being installed and 'delivery'
installed as a dependency of 'website_sale'.

closes odoo/upgrade#4713

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>