[FIX] gmail: circular dependencies and string refs
There are a few circular dependencies in the code which can result in
functions being called before they are actually defined in ES/6 import
order.

Lots of models just use the state to get a global property.
We replace these with access to the property directly, as the models
don't actually need to access the state otherwise.

The circular dependencies inside views are less problematic as the
view builders should only be called from the index in the first
place. And because we do not refer to any static data in the view builders.

As actions cannot accept anything but strings as parameter, the only fix
for these circular imports is to have a registry and reference the index
from the registry. However this introduce otherwise pointless indirection
so we leave things as they are for now.

----------------------------

Additionally, this commit replaces hard-coded strings for function
references inside actions with a reference to the name of the function
object. This helps avoid issues when functions are modified, ensure
the import order is correct and helps with avoiding these functions
from being marked as unused by linters and compilers.

task-3329768