[FIX] account/tests: use the latest CoA
Sometimes the generic CoA is not present, still we want to test with any
CoA available. Use the latest-installed one.

Also, even if the generic CoA is installed it may not be the right one
to test. For example when testing the upgrade saas-12.3->13 of
l10n_be_intrastat the test fails. It passes with this patch.

closes odoo/upgrade#4734

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>