[IMP] mrp_account_enterprise/saas~16.5: remove stat button
Since the Cost Analysis report is integrated into the MO Overview, there
is no longer the use of having the stat button in the MO form that leads
to it.
The report keeps existing however, as we still want to be able to create
Cost Analysis reports featuring multiple manufacturing orders.

See odoo/enterprise#40930

Part of task-3251506

closes odoo/upgrade#5210

Related: odoo/odoo#114536
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>