[FIX] account_asset/16.0: fix imported assets
To reproduce:
- Create a monthly asset in 15.0 with imported amount
(that began in the past and continues in the future)
- Upgrade
- Dispose it
=> The values of the disposal move are completely wrong

already_depreciated_amount_import was not taken into account.
It should be, as it completely changes the computation now.
We changed the values of the starting date to keep the same
depreciation computations, so old imported assets can be
modified/disposed/sold with right values.

opw-3315671

Part-of: odoo/upgrade#4765