[FIX] account/saas~13.4: fix long processing of draft statements
When there are too many draft statement lines the 48 bucket processing
loop is not fast enough. It seems this smallish bucket was chosen
because the query that fetches all the info to process returns a lot of
data when there are many draft lines (1 million + in current case).

```
2023-06-07 22:41:46,879 9501 INFO db_803773 odoo.upgrade.account.saas-13.4.end-09-payment-refactoring: Creating 1166300 moves for draft statement lines
```

In this patch we fetch the info with a server side cursor and increase
the bucket size. After some experimentation with 100K, 5K, and 1K sizes
it seems the faster processing is with 1K.

The total time to process the statements was around 19 hours.
```
2023-06-10 12:32:35,129 24 INFO 797035_afu_13.0_14.0 odoo.upgrade: [99.91%] 1166/1167 chunks processed in 19:09:12.132579 (total estimated time: 19:10:11.268199)
```

In the production server the DB is killed after 72 hours, the process
never ends.

upg-803773
opw-3295467

closes odoo/upgrade#4762

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>