[IMP] website/saas~16.4: remove data attributes on background
Steps to reproduce the problem:
- Add a "Text" snippet on the website.
- Change the background image with one of your own.
- Change the parallax to "Fixed".
- Save.
- Inspect the "Text" snippet.

-> There is a "data-snippet" and a "data-name" attribute on the
background image.

The goal of this commit is to remove the "data-snippet" and "data-name"
attributes on background images.

On the community PR, the problem is solved by filtering the data
attributes before transferring them to the new target at parallax
change.

See https://github.com/odoo/odoo/pull/123873

closes odoo/upgrade#4767

Related: odoo/odoo#123873
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>