[FIX] loyalty/saas~15.3.1.0 : recompute description for loyalty reward
From version 15 to version 16, we are merging 'coupon.reward' to
'loyalty.reward' so field 'reward_description' of coupon.reward model
is character type while in version 16, 'description' field of
'loyalty.reward' is compute field and with this description field's
value we are getting the value of product in 'disount_line_product_id'.

That's why, This PR will compute the value of description.

upg-783899, 840871, 840853, 840734

closes odoo/upgrade#4768

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>