[IMP] hr_contract_salary/saas~16.4: update record
'document' type of personal informations were not required on the salary
configurator, even when they were set as required on the personal info.

Due to this, prior, hr_contract_salary_personal_info_image_1920 was not
required on the salary configurator. In order to not change the bihaviour
and tours, we set it to not required on the personal info.

Enterprise PR: odoo/enterprise#42619

task-3266175

closes odoo/upgrade#4783

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>