[FIX] loyalty/saas~15.3.1.0: Remapping Translation
During upgrade loyalty_program and coupon_program are merged but the
translation is not migrated properly because the script was prepared
to handle all the diffrent possiblity of modules installed. So,
when both coupon_program and loyalty_program are installed merge_model
method is used which will delete all coupon_program's transltion as the
we have used remove_model function in the end of merge_model.
So to solve this issue the records for the coupon_program's translation
are detached and a _remap_translation function is used to remap and attach
the records to loyalty_program accordingly.

req-797602
opw-3325114

closes odoo/upgrade#4784

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>