[FIX] account/0.0.0: do not compute release_to_pay
A recompute of many `account.move`s is triggered manually from
migrations/account/0.0.0/end-fix_is_matched.py, which in turn
computes `release_to_pay` that is a field protected from recomputation
in case its corresponding move is already reconciled.

We narrow down the set of payments to run the recompute on to be more
in line with what the original tech support script does.

closes odoo/upgrade#4801

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>