[FIX] project{_todo}/saas~16.4: migrate db with note from 16 to 16.4
Before this commit, when a db with just note installed is migrated
from 16 to saas~16.4, the migration failed because the `is_closed`
field does not exist in the `project_task`. The reason is because
the project module will be installed when `project_todo` will be installed
to replace note module during the migration. It means the project module
installed will be in version 16.4 and in that version `is_closed`
field is removed in `project.task` model and so during the migration
of `project_todo`, the `is_closed` used in SQL query cannot be found.

This commit fixes the issue by using the `is_closed` column in
`project_task` only if the target version is lower than saas~16.4.

closes odoo/upgrade#4819

X-original-commit: fbc6943eb20caee375c53154e0009e22f8467439
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>