[FIX] base/saas~16.3: exception to note removal
the notes module is deleted if it is detected as
unused to avoid offering another module to the
free databases.

We also need to check that the client is not
using the activity type defined by the module.
If this is the case, we delete the xmlid to
preserve the client's data and not prevent the
module from being deleted.

upg-857540

closes odoo/upgrade#4817

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>