[FIX] sale_project/saas~16.3: filter projects to make billable
Before this commit, when a db is migrated to saas~16.3, all projects
with partner set or at least one of their tasks with partner set will
be billable after the migration to be sure the user will continue
to see the partner field in those projects and linked tasks.
However, according to odoo/enterprise#36459, the fsm project does not
need to be billable to have the partner field visible on project and
tasks and so it is really need to make them billable during the
migration.

This commit checks if `industry_fsm_sale` is installed if yes, then
filter to projects to make billable to only update non fsm projects.

Community PR: odoo/odoo#125622
Enterprise PR: odoo/enterprise#43352

closes odoo/upgrade#4796

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>