[FIX] account_asset/tests: Fiscal year last day localization
In India, the last month of the fiscal year is not December, but March.
It causes issues in the test, because the depreciation entries are posted
following this date.
But adapting to the fiscal year would change all values expected,
while lowering the ease of understanding them.
So, we impose a fiscal year date in the prepare section.

Linked to runbot error
https://runbot.odoo.com/web/#id=22385&view_type=form&model=runbot.build.error&menu_id=405&cids=1

closes odoo/upgrade#4870

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>