[FIX] base/saas~16.{1,4}.1.3: l10n_pe_reports renamed without script
A module named l10n_pe_reports was introduced, and during the
fw-port, was renamed l10n_pe_reports_book, because a module with
this name already existed.

In master, both modules are merged.
We will rename it in 16.1.

In 16.4, we have to rename the module if only one module is present,
meaning we're migrating from 16 to 16.4, and the ORM will
just add fields and views in the module.
If both are present, we're simply merging them.

Related: https://github.com/odoo/enterprise/pull/43354
Related to runbot error:
https://runbot.odoo.com/web/#id=21742&view_type=form&model=runbot.build.error&menu_id=405&cids=1

closes odoo/upgrade#4872

Related: odoo/enterprise#43354
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>