[IMP] {helpdesk_}timesheet{_grid}/saas~18.3: create timesheets for timer
This commit generates the timesheets for the existing timers linked to
ticket or task and stopped those timers because now when a timer will
be started in task or ticket, a timesheet is generated (before the
timesheet was generated when the timer is stopped). Those timers are
stopped because it is possible a timer is existed for a same user in
another timesheet. To avoid having 2 timers in the same model for the
same user, those timers are removed after creating the timesheets.

task-2276015

Enterprise PR: odoo/enterprise#16084

closes odoo/upgrade#2208

Related: odoo/upgrade-util#51
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>