[IMP] approvals/saas~16.5: fix access error
Previously:
Simple user didn't see the approver lines that they were assigned to,
unless it was their own request.

In the enterprise commit:

We give read access to the simple user to the approver lines that they
are assigned to (they are set as user_id).
As a result, we need to update the rule.

Enterprise PR: odoo/enterprise#43661

task - 3339428

closes odoo/upgrade#4890

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>