[IMP] hr_timesheet_attendance/saas~16.4: base report on employees
Related:https://github.com/odoo/odoo/pull/127492
The report is currently base on timesheets/attendances.user_id.employee_id.
As a result, some data weren't taken into account.
As both models have employee_id, there is no reason not to use it.

Also, we had a view in this module that was not used
and didn't seem to make any sense, so we delete it here.

task-3251675

closes odoo/upgrade#4900

Related: odoo/odoo#127492
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>