[FIX] account_asset/16.0: update prorata date computation
The computation of the prorata date misses the cases where the
total value is not exactly a multiple of the periodic values, and
when the remainder of original - previously depreciated is very
small, which using the previous equation makes it seem like the
depreciation started thousands of years ago.
It could be better to try to get an estimate of a periodic
payment value, then get the number of payments by dividing
the previously depreciated value by the periodic rate.

closes odoo/upgrade#4921

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>