[IMP] website_{,slides_}forum/saas~16.4: redesign forum
Remove views and field obsolete with Forum's redesign.

forum image from course
-----------------------

A new compute is added for the forum Image to use
the images of the related course where applicable.
Therefore, we clone the relevant rows if no image was
already set on the forum.

bump date replaced by last_activity date
----------------------------------------

The bump date is deleted and replaced by a new field: last_activity date.
As this field will be used as the default sorting of the forum posts,
it cannot be null.
Instead of implementing a complex operation to impose the new
last_activity computation and alter the existing order of posts,
we decided to set the last_activity date to the write date and let
the records update themselves with usage thanks to the
functionalities implemented in the associated community PR.

See odoo/odoo#123709
See odoo/enterprise#42388

Task-3349374
Task-3083657
Task-3316555

closes odoo/upgrade#4764

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Lopes Marc (loma) <loma@odoo.com>
Co-authored-by: Florian Charlier <flch@odoo.com>