[FIX] sale_subscription/16.0.1.1: Remove deprecated sequence
In V16, the sale_subscription model has been merged into the sale_order model
and the sequence of the sale_subscription model has been removed,
but during migration, the sequence 'seq_sale_subscription' is not being
removed. It creates confusion in the customer's mind that the sequence
is exist but not working.

opw-3346051
upg-1052319

closes odoo/upgrade#4961

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>