[IMP] timer,*/saas~18.3: uniformize the using of timer
* = helpdesk_timesheet,timesheet_grid

The timer used in the Helpdesk, Project and Timesheet app was linked to 3
different models (`helpdesk.ticket`, `project.task` and
`account.analytic.line`), the problem with that is when the timer is
launched on ticket or a task, the user cannot stops the timer in the
Timesheet app.

The goal is the timer will be linked to a timesheet and ticket/task (if
the timesheet linked is linked to a ticket or a task). By doing that,
the user will be able to stop the timer in the helpdesk/Project app and
Timesheet app.

This commit migrates the changes to be able to align the timer in
tickets/tasks with the timer header in Timesheet app.
That is:
- the `helpdesk.ticket` and `project.task` will inherit to
`timer.parent.mixin` instead of `timer.mixin`.
- 2 fields are added in `timer.timer` model to be able to link the
ticket or the task on the timer.
- The `project.task.create.timesheet` and
`helpdesk.ticket.create.timesheet` wizards are removed because those
wizards were used when the timer is stopped in the form view of task
(respectivetly, ticket) to create a timesheet after the changes in
odoo/enterprise#65414 , the timesheet will be generated when the timer
will be started and so those wizards will not be useful anymore except
if we want to add confirmation for the user.

task-2276015

Enterprise PR: odoo/enterprise#16084

Part-of: odoo/upgrade#2208
Related: odoo/upgrade-util#51
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>