[FIX] account_(peppol,edi_proxy_client)/saas~16.5.1.0: remove fields
In `account_peppol`:
- `peppol_migration_key` field has to be stored on `res.company`
instead of `account_edi_proxy_client.user`.
Currently, this (optional) value is actually saved
in config params as a workaround,
with a dynamic key `account_peppol.migration_key_{company_id}`.
There can be one migration key per company and it is only
created in a specific case, however it is still necessary to copy
the migration keys from config parameters
to res_company, deleting the corresponding config parameters.
The `peppol_migration_key` on the `account_edi_proxy_client.user` is not used,
so can be safely deleted.
- due to a changed spec `account_peppol_attachment_ids` is not needed anymore.
It is not actually used, so can be just removed.
- `account.move.send` now has `enable_peppol` field and `peppol_proxy_state`
is no longer needed.

In `account_edi_proxy_client`:
- SQL constraints `unique_edi_identification` and `unique_company_proxy`
should only consider active edi users,
i.e. archived users should not be included in the constraint.

closes odoo/upgrade#4980

See: odoo/odoo#129595
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>