[IMP] pos_self_order/saas~16.5: product attributes
When selecting an attribute value in POS, what we store is the `description`,
which is a string that represents the selected attribute values.
Ex: selecting `Size: M` and `Material: Leather` will result in the description:
`(L, Leather)`.

This choice does not lead to a logical API for dealing with product attibutes.

In this pr we remove the `selected_attributes` field in which we used to
store the `string` description and replace it with the `attribute_Value_ids`
field which is a `Many2many` field linking the orderline with the specific
values chosen.

closes odoo/upgrade#4992

Odoo: odoo/odoo#126398
Task: 3378533
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>