[FIX] repair/16.4.1.0: fix date tolerance in test
When the upgrading process take place on two different days,
the verification of the field 'under_warranty' based on the
current date may become wrong.

This is due to a loss of information in the refactoring process :
 - in old model, the field 'guarantee_limit' was a date,
   such that it had a large data scope for comparison.
 - in new model, the replacing field 'under_warranty' is a boolean.

That loss of information is intended, this fix only affect the test
methods, such that the migration script itself remains fully operative.

Error :
>    [...]/test_16_4_repair_refactor.py", line 331, in check
>        self.assertTrue(ro.under_warranty)
>    AssertionError: False is not true

Prerequisites :
 - Odoo (Enterprise or Community) database in version < 16.4
 - Repair module installed

Steps to reproduce :
 - On day D, launch the database with:
    - a correct upgrade-path
    - -u repair
    - --test-tags upgrade.test_prepare
 - On day D+1
    - update the database to version 16.4
    - --test-tags upgrade.test_check

Community : odoo/odoo#106911
Enterprise : odoo/enterprise#36088

closes odoo/upgrade#4994

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>