[FIX] base/saas~16.2: fix iap_extract deps
The module iap_extract was auto_installed when mail was installed
but now the dependencies include iap module. For databases having
the module installed with the old dependency, the module should
be uninstalled.
The affected versions are saas~16.2, saas~16.3, and saas~16.4 given
that the module was introduced in saas~16.1 and that the highest
version available is saas~16.3.
https://github.com/odoo/enterprise/pull/44757

closes odoo/upgrade#4995

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>