[IMP] website_slides_survey/saas~17.2: add certfied stat btn and filter
We added a boolean field on the `slide_channel_partner` model to keep track
of which course attendee is certified or not. This new field needs to be
updated with data from `slide_slide_partner`.

Aslo, replacing the Finished button with a Certified button in slide_channel
kanban view maps is no longer appropriate.
The slide_channel kanban view override is no longer useful and has been
removed.

see odoo/odoo#130525

task-3412724

closes odoo/upgrade#5006

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>