[FIX] l10n_be_hr_payroll,l10n_be_hr_contract_salary/saas~16.5: fixup
In the PR odoo/enterprise#41397 the field 'has_bicycle' was moved from
employee to contract. But the modularity was not respected.
It was removed in l10n_be_hr_payroll and added in
l10n_be_hr_contract_salary.
Thus the corresponding upgrade script odoo/upgrade#4717 caused issues.
In particular it raised- hr.employee.has_bicycle has not been removed.

It this commit, we fix that.

enterprise PR - odoo/enterprise#45076

closes odoo/upgrade#5015

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>