[FIX] base/saas~16.3.1.3: set category of mail_activity_type
From version saas~16.2 to saas~16.3, Uninstalling 'Note' module if 'note.note'
model don't have any record.
In case, customer's database  don't have any record of 'note.note' but
'reminder' mail_activity_type (that comes from 'Note' module) and that is used
in mail_activity of any record then it will raise an error of "activity_category
field have wrong value of mail_activity model" due to the fact that
activity_category is non-store and related field to category field of
mail_activity_type and category of reminder type mail activity is set as null
due to during upgrade we deleted ir_model_data record of reminder type and as
we uninstalling a 'note' module. The category of that is also removed .
To resolve this issue, I have updated category of reminder mail_activity_type to
default for that case.

upg-1053393

closes odoo/upgrade#5014

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>