[IMP] l10n_mx_edi/saas~16.5.0.3: simplify l10n_mx_edi_certificate
Currently, there is a m2m relationship between res_company
and l10n_mx_edi_certificate
(`l10n_mx_edi_certificate_res_company_rel`). This is going to change in 16.5,
where we add a required m2o company_id field on the certificate.

Even though certificates are not supposed to be shared between companies,
the current design allows for it.
So it is necessary to keep those certificates during upgrade.
The idea is to copy the certificate data and insert a new row
for every company that shares that certificate.
First, the existing rows are updated with the company_id.
Then we insert duplicate certificates for every other company.

See:
odoo/enterprise#44971

closes odoo/upgrade#5011

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>
Co-authored-by: dbkosky dako@odoo.com