[IMP] worksheet/saas~17.2: replace company_ids by a m2o
This commit replaces the company_ids field defined in worksheet.template by a
many2one field called company_id. The idea of that changes is to be coherent
with the other apps ( as the norm is to have a m2o for multi-company).

Ent PR: https://github.com/odoo/enterprise/pull/45227

task-3450204

closes odoo/upgrade#5048

Related: odoo/enterprise#45227
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>