[FIX] account/16.0: set reconciled 0 amount invoices
Invoices that have already been paid in full
show up on the contact's Follow Up Report.
so Invoices which has 0 residual amount also show
on followup report which is wrong. reason
in version 15, when we create invoice and
there is zero residual amount then it will not set
`reconciled` true for receivable account.
and that will create issue as we are going to
displayed those Journal items which are
unreconciled and which related account type
is `asset_receivable`.
see:
https://github.com/odoo/enterprise/blob/4bdae5a59d758df0837c80a8e5ab09dad81eca68/account_followup/models/res_partner.py#L180

In older version if we create this kind of entry
with zero residual amount, it was not set `reconciled`
true for `receivable` account type journal item.
because compute field code for `reconciled` is different
it checked `account_partial_reconcile` table entry
to set it true.
https://github.com/odoo/odoo/blob/95fccaba4555f2be6712154a84a143f589a3f54c/addons/account/models/account_move.py#L4491

and from version 16 , it has been changed:
https://github.com/odoo/odoo/blob/2b9d90484d0b4058bfd39d83a8dceca5047a3c27/addons/account/models/account_move_line.py#L755

upg-1056019
upg-856142
upg-789667
upg-859645

unicorn task: 3452682

closes odoo/upgrade#5026

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>